moving interfaces to adapters layer
