+++ src/test/java/com/samir/uberweal/application/command/handlers/RiderPreAuthorizeTest.java
@@ -5,8 +5,8 @@ import com.samir.uberweal.application.query.handler.PreAuthorizeQueryHandler;
 import com.samir.uberweal.domain.entities.Rider;
 import com.samir.uberweal.domain.exceptions.InsufficientFundsException;
 import com.samir.uberweal.domain.exceptions.RiderNotFoundException;
-import com.samir.uberweal.domain.gateways.RiderDsGateway;
-import com.samir.uberweal.domain.gateways.stubs.RiderDsGatewayStub;
+import com.samir.uberweal.adapters.gateways.RiderDsGateway;
+import com.samir.uberweal.adapters.gateways.stubs.RiderDsGatewayStub;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
